backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
redis==5.0.1
celery==5.3.4
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
Pillow==10.1.0
opencv-python==4.8.1
openai==1.6.1
anthropic==0.8.1
pytest==7.4.3
pytest-asyncio==0.21.1
